Configuration information section
In configuration information section the command specified to check Activity master has to be modified to (Get-AdfsFarmInformation).FarmRoles. If we execute the command as Get-AdfsFarmInformation.FarmRoles then it errors stating 

Get-AdfsFarmInformation.FarmRoles : The term 'Get-AdfsFarmInformation.FarmRoles' is not recognized as the name of a cmdlet, function, script file, or operable program. Check the spelling of the name, or if a path was included,
verify that the path is correct and try again.
At line:1 char:1
+ Get-AdfsFarmInformation.FarmRoles
+ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    + CategoryInfo          : ObjectNotFound: (Get-AdfsFarmInformation.FarmRoles:String) [], CommandNotFoundException
    + FullyQualifiedErrorId : CommandNotFoundException